requirements.txt
numpy
pandas
tqdm
ffmpeg-python
pydub
librosa
openai-whisper
scipy
